cambios despues de refactoring y orden

+++ src/main/java/com/app/inmuebles/formasMenu/FormasMenuServiceImpl.java
@@ -5,7 +5,7 @@
  */
 package com.app.inmuebles.formasMenu;
 
-import com.app.inmuebles.formasMenu.FormasMenu;
+import com.app.inmuebles.util.Mensaje;
 import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -19,20 +19,27 @@ public class FormasMenuServiceImpl implements FormasMenuService{
 
     @Autowired
     private FormasMenuDAO formasMenuDAO;
+    private Mensaje msg;
 
     @Override
     public List<FormasMenu> listAll() {
-        return formasMenuDAO.getRegistros();
+        return formasMenuDAO.getRecords();
     }
 
     @Override
     public List<FormasMenu> listAllFathers() {
-        return formasMenuDAO.getRegistrosp();
+        return formasMenuDAO.getRecordsFather();
     }
 
     @Override
-    public int addFormasMenu(FormasMenu formasMenu) {
-        return formasMenuDAO.addFormaMenu(formasMenu);
+    public Mensaje addFormasMenu(FormasMenu formasMenu) {
+        int valor =formasMenuDAO.addFormaMenu(formasMenu);
+        if (valor >= 1) {
+            msg = new Mensaje("Agregado correctamente", 1);
+        } else {
+            msg = new Mensaje("No se pudo agregar", 2);
+        }
+        return msg;
     }
 
     @Override
@@ -41,18 +48,54 @@ public class FormasMenuServiceImpl implements FormasMenuService{
     }
 
     @Override
-    public int editFormasMenu(FormasMenu formasMenu) {
-        return formasMenuDAO.editFormaMenu(formasMenu);
+    public Mensaje editFormasMenu(FormasMenu formasMenu) {
+        boolean existe = formasMenuDAO.existsHijos(formasMenu.getNoFormaMenu());
+        if (existe) {
+            if (formasMenu.getNoFormaPadre() != 0) {
+                msg = new Mensaje("No se pudo editar", 2);
+            } else {
+                int valor = formasMenuDAO.editFormaMenu(formasMenu);
+                if (valor >= 1) {
+                    msg = new Mensaje("Editado correctamente", 1);
+                } else {
+                    msg = new Mensaje("No se pudo editar", 2);
+                }
+            }
+        } else {
+            int valor = formasMenuDAO.editFormaMenu(formasMenu);
+            if (valor >= 1) {
+                msg = new Mensaje("Editado correctamente", 1);
+            } else {
+                msg = new Mensaje("No se pudo editar", 2);
+            }
+        }
+        return msg;
     }
 
     @Override
-    public int deleteFormasMenu(int id) {
-        return formasMenuDAO.deleteFormaMenu(id);
+    public Mensaje deleteFormasMenu(int id) {
+        boolean existe = formasMenuDAO.existsHijos(id);
+        if (existe) {
+            msg = new Mensaje("No se pudo eliminar por que tiene pantallas asociadas", 2);
+        } else {
+            boolean existe2 = formasMenuDAO.existsRolFormas(id);
+            if (existe2) {
+                msg = new Mensaje("No se pudo eliminar por que la pantalla esta asignada a un Rol", 2);
+            } else {
+                int valor = formasMenuDAO.deleteFormaMenu(id);
+                if (valor >= 1) {
+                    msg = new Mensaje("Eliminado correctamente", 1);
+                } else {
+                    msg = new Mensaje("No se pudo eliminar", 2);
+                }
+            }
+        }
+        return msg;
     }
 
     @Override
-    public List<String> getPermisoPantalla(int noUsuario) {
-        return formasMenuDAO.getPermisoPantalla(noUsuario);
+    public List<String> getPermissionToPages(int noUsuario) {
+        return formasMenuDAO.getPermissionToPages(noUsuario);
     }
 
     @Override
